--- a/server/channels/db/migrations/postgres/000135_sidebarchannels_categoryid.down.sql
+++ b/server/channels/db/migrations/postgres/000135_sidebarchannels_categoryid.down.sql
@@ -1,0 +1,2 @@
+-- morph:nontransactional
+DROP INDEX CONCURRENTLY IF EXISTS idx_sidebarchannels_categoryid;