--- a/server/channels/db/migrations/postgres/000135_sidebarchannels_categoryid.up.sql
+++ b/server/channels/db/migrations/postgres/000135_sidebarchannels_categoryid.up.sql
@@ -1,0 +1,2 @@
+-- morph:nontransactional
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_sidebarchannels_categoryid ON sidebarchannels(categoryid);